Ensure supporting files are included in source distribution
These supporting files should be included in the source
distribution shipped to PyPI. Most critically, the LICENSE
ought to be shipped with the source.

Also removes the code in setup.py tolerating absence of README.md.
The reason it was missing for older python/setuptools was that
README.md isn't implicitly included in the source dist for those
old versions.